Refactor metaworkflow app
Refactored the metaworkflow app, extracting methods in the
entrypoint into their own module
